Move Model classes
Harmonize directory structure with other bounded context repos.
