+++ config/DoctrineClassMapping/WMDE.Fundraising.AddressChange.Domain.Model.Address.dcm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="WMDE\Fundraising\AddressChange\Entities\Address" table="address">
+  <entity name="WMDE\Fundraising\AddressChange\Domain\Model\Address" table="address">
     <id name="id" type="integer" column="id">
       <generator strategy="IDENTITY"/>
     </id>
