+++ config/DoctrineClassMapping/WMDE.Fundraising.AddressChange.Domain.Model.AddressChange.dcm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="WMDE\Fundraising\AddressChange\Entities\AddressChange" table="address_change">
+  <entity name="WMDE\Fundraising\AddressChange\Domain\Model\AddressChange" table="address_change">
     <indexes>
       <index name="ac_export_date" columns="export_date"/>
       <index name="IDX_7B0E7B9FF5B7AF75" columns="address_id"/>
